+++ src/containers/GameMenuContainer.jsx
@@ -5,11 +5,11 @@ import GameMenuView from '../views/GameMenuView';
 
 class GameMenuContainer extends PureComponent {
 
-  getMenuItems(props) {
+  getButtons(props) {
     const {actions, state} = props;
     const togglePersp = () => actions.GameMenu.togglePerspective(!state.GameBoard.hasPerspective);
-    const reloadLevel = () => actions.GameMenu.updateLevel(state.GameBoard.level);
-    const incrLevel = () => actions.GameMenu.updateLevel(state.GameBoard.level + 1);
+    const reloadLevel = () => actions.GameMenu.reloadLevel();
+    const incrLevel = () => actions.GameMenu.updateLevel(Math.min(12, state.GameBoard.level + 1));
     const decrLevel = () => actions.GameMenu.updateLevel(Math.max(1, state.GameBoard.level - 1));
     const incrLightRadius = () => actions.Player.updateLightRadius(state.Entity.Player.spawns[0].lightRadius + 1);
     const decrLightRadius = () => actions.Player.updateLightRadius(state.Entity.Player.spawns[0].lightRadius - 1);
@@ -23,11 +23,9 @@ class GameMenuContainer extends PureComponent {
       title: `level: ${state.GameBoard.level}`,
       buttons: [
         <button key="ReloadLevel" onClick={reloadLevel}>reload</button>,
-        <button key="DecrLevel" onClick={decrLevel}>-</button>,
-        <button key="IncrLevel" onClick={incrLevel}>+</button>
+        <button key="DecrLevel" disabled={state.GameBoard.level === 1} onClick={decrLevel}>-</button>,
+        <button key="IncrLevel" disabled={state.GameBoard.level === 12} onClick={incrLevel}>+</button>
       ]
-    }, {
-      title: `number of tiles: ${Math.pow(2, this.props.state.GameBoard.level)}`
     },{
       title: `lightradius: ${state.Entity.Player.spawns[0].lightRadius}`,
       buttons: [
@@ -37,10 +35,23 @@ class GameMenuContainer extends PureComponent {
     }]
   }
 
+  getInfo(props) {
+    const { state } = props;
+
+    return [{
+      title: `position: ${state.Entity.Player.spawns[0].position.tileId + 1}`
+    }, {
+      title: `room: ${state.Entity.Player.spawns[0].position.roomId}`
+    }, {
+      title: `tiles: ${Math.pow(2, state.GameBoard.level)}`
+    }]
+  }
+
   render() {
     return (
       <GameMenuView>
-        {this.getMenuItems(this.props)}
+        {this.getButtons(this.props)}
+        {this.getInfo(this.props)}
       </GameMenuView>
     );
   }