+++ src/views/GameMenuView.jsx
@@ -3,15 +3,18 @@ import './GameMenuView.scss';
 
 export default props => {
   return <div className='game-menu'>
-    {
-      props.children.map((menu, index) => {
-        return (
-          <div className="menu-block" key={ index }>
-            <h2>{ menu.title }</h2>
-            { menu.buttons }
-          </div>
-        )
-      })
-    }
+    <div className="menu-block">{ props.children[0].map((menu, index) => (
+      <div className="item" key={ index }>
+        <h2>{ menu.title }</h2>
+        { menu.buttons }
+      </div>
+    )) }</div>
+    <div className="menu-block">{ props.children[1].map((menu, index) => (
+      <div className="menu-item" key={ index }>
+        <h2>{ menu.title }</h2>
+        { menu.buttons }
+      </div>
+    )) }</div>
+
   </div>;
 };